backend/pyproject.toml
[project]
name = 'yatzy-backend'
version = '0.1.0'
requires-python = '>=3.14'
dependencies = ['fastapi[standard]', 'aiomysql', 'pydantic-settings', 'cryptography', 'PyJWT', 'yatzy-rules']

[tool.uv.sources]
yatzy-rules = {path = '../shared'}

[project.optional-dependencies]
dev = [
    'pytest',
    'pytest-asyncio',
    'httpx',
    'ruff',
    'ty',
    'pytest-cov',
    'bandit',
    "pytest-shard>=0.1.2",
]

[tool.ruff]
indent-width = 2

[tool.ruff.format]
quote-style = 'single'

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'

[tool.bandit]
skips = ['B101', 'B311']

[tool.coverage.paths]
source = ['src/app/', '/workspace/backend/src/app/']

[tool.coverage.report]
fail_under = 90

[tool.pytest.ini_options]
asyncio_mode = 'auto'
testpaths = ['src/tests']
pythonpath = ['src']
addopts = ''
